+++ src/language-server/__tests__/kerml/core/classifiers.test.ts
@@ -27,6 +27,7 @@ describe.each(["specializes", "conjugates"])(
                 {},
                 {
                     diagnostics: [defaultLinkingErrorTo("A")],
+                    buildOptions: { validationChecks: "all", standardLibrary: "none" },
                 }
             );
         });